Add new App.js, Header and Drawer component (LeftMenu)
